divide foloer & create storyboard
