--- nanuri-app/nanuri-app/Controller/ViewController.swift
@@ -1,21 +1,0 @@
-//
-//  ViewController.swift
-//  nanuri-app
-//
-//  Created by minimani on 2021/11/16.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-        print("gitTest")
-        print("adddevelopbranch")
-    }
-
-
-}
-
