--- nanuri-app/nanuri-app/Model/Test.swift
@@ -1,8 +1,0 @@
-//
-//  Test.swift
-//  nanuri-app
-//
-//  Created by minimani on 2021/11/16.
-//
-
-import Foundation
