--- nanuri-app/nanuri-app/Utils/Base.lproj/Main.storyboard
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="19455" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="Gdo-yD-cFx">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19454"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--Tab Bar Controller-->
-        <scene sceneID="fan-Gi-phR">
-            <objects>
-                <tabBarController id="Gdo-yD-cFx" sceneMemberID="viewController">
-                    <navigationItem key="navigationItem" id="LoU-V7-a7C"/>
-                    <tabBar key="tabBar" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="1I7-oD-fh0">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="49"/>
-                        <autoresizingMask key="autoresizingMask"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                    </tabBar>
-                    <connections>
-                        <segue destination="eUJ-GH-F9F" kind="relationship" relationship="viewControllers" id="itG-g3-xWL"/>
-                    </connections>
-                </tabBarController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="rFz-PS-LWt" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="667" y="-141"/>
-        </scene>
-        <!--UINavigationController-wN3-ez-dQs-->
-        <scene sceneID="6Fl-Ds-kTj">
-            <objects>
-                <viewControllerPlaceholder storyboardName="Home" referencedIdentifier="UINavigationController-wN3-ez-dQs" id="eUJ-GH-F9F" sceneMemberID="viewController">
-                    <tabBarItem key="tabBarItem" title="Item" id="XdT-XE-4C7"/>
-                </viewControllerPlaceholder>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="Q4g-Qx-nQy" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="420" y="289"/>
-        </scene>
-    </scenes>
-</document>
