+++ nanuri-app/nanuri-app/View/Base.lproj/Main.storyboard
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="19455" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="Ppz-bq-7ka">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19454"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <scenes>
+        <!--Tab Bar Controller-->
+        <scene sceneID="f9r-9c-wsy">
+            <objects>
+                <tabBarController id="Ppz-bq-7ka" sceneMemberID="viewController">
+                    <navigationItem key="navigationItem" id="b24-uv-72K"/>
+                    <tabBar key="tabBar" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="eW2-Za-wzr">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="49"/>
+                        <autoresizingMask key="autoresizingMask"/>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                    </tabBar>
+                    <connections>
+                        <segue destination="0XO-UF-k3a" kind="relationship" relationship="viewControllers" id="ZQ8-bg-7BZ"/>
+                        <segue destination="GQ5-OO-QTa" kind="relationship" relationship="viewControllers" id="RdU-Bm-GGS"/>
+                        <segue destination="zKP-3M-RE8" kind="relationship" relationship="viewControllers" id="05R-H3-AYp"/>
+                    </connections>
+                </tabBarController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="Al1-Mf-0oR" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="667" y="-141"/>
+        </scene>
+        <!--List-->
+        <scene sceneID="SoW-Pb-TfL">
+            <objects>
+                <viewControllerPlaceholder storyboardIdentifier="UINavigationController-6TN-Pp-m51" storyboardName="List" id="GQ5-OO-QTa" sceneMemberID="viewController">
+                    <tabBarItem key="tabBarItem" title="Item" image="homekit" catalog="system" id="2gI-tJ-LmE"/>
+                </viewControllerPlaceholder>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="URt-IN-JAb" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="667" y="289"/>
+        </scene>
+        <!--Home-->
+        <scene sceneID="0Lp-yG-8p4">
+            <objects>
+                <viewControllerPlaceholder storyboardIdentifier="UINavigationController-wN3-ez-dQs" storyboardName="Home" id="0XO-UF-k3a" sceneMemberID="viewController">
+                    <tabBarItem key="tabBarItem" title="Item" id="ztz-M1-D92"/>
+                </viewControllerPlaceholder>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="SDQ-ce-xNg" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="379" y="289"/>
+        </scene>
+        <!--MyPage-->
+        <scene sceneID="y6g-mo-ovm">
+            <objects>
+                <viewControllerPlaceholder storyboardName="MyPage" id="zKP-3M-RE8" sceneMemberID="viewController">
+                    <tabBarItem key="tabBarItem" title="Item" id="Jld-pt-Wy2"/>
+                </viewControllerPlaceholder>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="KAe-bQ-5U0" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="954" y="288"/>
+        </scene>
+    </scenes>
+    <resources>
+        <image name="homekit" catalog="system" width="128" height="106"/>
+    </resources>
+</document>
